+++ Final-Exam/src/main/java/com/FinalExam/repository/RecordsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RecordsRepository extends JpaRepository<Record, Long> {
+public interface RecordsRepository extends JpaRepository<Record, Integer> {
 }
