+++ Final-Exam/src/main/java/com/FinalExam/repository/TeamsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TeamsRepository extends JpaRepository<Team, Long> {
+public interface TeamsRepository extends JpaRepository<Team, Integer> {
     Team getTeamById(int id);
 }
